Check in compile-time for missing commands
During compile-time ensure that the users won't be able to
get the arguments or check for presence for a command that
has not been provided earlier in a tuple  when parsing the
commands.

Warning: This is currently not completely fool-proof as we
don't take into consideration the 'id' of the command.
This means that the user may forget to pass one command that
happens to have the same arguments as another still.